Getting ready for rewriting
